+++ node-test/bindings/global.mli
@@ -7,7 +7,7 @@ val interval_id_to_js: interval_id -> Ojs.t
 val interval_id_of_js: Ojs.t -> interval_id
 
 
-val set_interval: (unit -> unit) -> int -> interval_id
-val set_timeout: (unit -> unit) -> int -> timeout_id
-val clear_timeout: timeout_id -> unit
-val clear_interval: interval_id -> unit
+val set_interval: (unit -> unit) -> int -> interval_id [@@js.global]
+val set_timeout: (unit -> unit) -> int -> timeout_id [@@js.global]
+val clear_timeout: timeout_id -> unit [@@js.global]
+val clear_interval: interval_id -> unit [@@js.global]